Markets and economies can now completely eschew consumptionManager when trying to fulfill demand

+++ Assets/Company/CompanyBuilder.cs
@@ -20,7 +20,7 @@ public class CompanyBuilder<T> where T : Company
         companyToReturn.SetStrategy(behaviourStrategy);
         return this;
     }
-    public CompanyBuilder<T> WithInitialCash( int initialCash)
+    public CompanyBuilder<T> WithInitialCash(decimal initialCash)
     {
         companyToReturn.SetCash(initialCash);
         return this;
